# Image shard generation