@william-silversmith 2020-08-11
https://github.com/seung-lab/cloud-volume/issues/264 is related to eeb55a2