@william-silversmith 2020-08-11
If downloading entire shards, I really recommend just downloading the file and using `reader.disassemble_shard`.