@nkemnitz 2020-08-11
> If downloading entire shards, I really recommend just downloading the file and using `reader.disassemble_shard`.

You mean I should check if bounding box envelops an entire shard and have a special code path to optimize for that? Or that the user should be responsible for downloading and disassembling shards?