@william-silversmith 2020-08-11
You might want to run a test to see if there's a significant difference in
speed. One advantage of downloading the whole file is I can compute the md5
so that's nice.

On Tue, Aug 11, 2020, 12:28 PM Nico Kemnitz <notifications@github.com>
wrote:

> If downloading entire shards, I really recommend just downloading the file
> and using reader.disassemble_shard.
>
> You mean I should check if bounding box envelops an entire shard and have
> a special code path to optimize for that? Or that the user should be
> responsible for downloading and disassembling shards?
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/seung-lab/cloud-volume/pull/371#issuecomment-672070830>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AATGQSJJPJA5LHC2ZJB2UBDSAFWRLANCNFSM4PO6O3KA>
> .
>