@spikeliu 2021-08-05
Combined with the information from #221, I am little confused now, does cloud-volume support generating a volume in sharded format and upload it, or it only supports shard reads? I looked at the make_shard method, and it seems to generates a sharded file in memory, so at least it limits how large the file can be (can not beyond machine's memory). Also, it looks like there is no flag or method exposed to users for specifying sharded format when uploading the volume.