@william-silversmith 2021-08-05
This is kind of a dumb state of affairs, but we currently have the ability to create a sharded volume from an unsharded one easily using Igneous, but we don't have an easy way to create one from scratch with CloudVolume other than via `make_shard` + CloudFiles upload. I haven't fully thought it through how one might do that, but the main reason is that creating new datasets from e.g. tiffs is fairly idiosyncratic. The shard specification dictates how big an uploaded image shard is supposed to be. Igneous now can generate a reasonable sharding specification automatically, so perhaps we could just tell the user what size image blocks to use...