@william-silversmith 2021-08-05
Here's how we generate new sharded volumes:

https://github.com/seung-lab/igneous/blob/master/igneous/tasks/image.py#L472-L531
https://github.com/seung-lab/igneous/blob/master/igneous/task_creation/image.py#L387-L488